Refactor: 컬럼 폴더 구조 리팩토링
